+++ packages/widget_test_harness/lib/src/ui_harness_helpers/callback_base_extensions.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:widget_test_harness/src/given_when_then.dart';
+
+import '../test_harness.dart';
+
+extension CallbackClassBaseWidgetTester
+    on GWTBase<WidgetTestHarness> {
+  @protected
+  WidgetTester get tester => harness.tester;
+}