+++ packages/widget_test_harness/lib/src/ui_harness_helpers/given.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:widget_test_harness/src/ui_harness_helpers/callback_base_extensions.dart';
 
 import '../given_when_then.dart';
 import '../test_harness.dart';
 
 extension CommonGivenWidgetTestHelpers on Given<WidgetTestHarness> {
-  @protected
-  WidgetTester get tester => harness.tester;
   /// passthrough for [WidgetTester.pumpWidget]
   Future<void> pumpWidget(
     Widget widget, [